requirements.txt
# Core Python data handling
pandas>=1.5.0

langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-ollama>=0.1.0


# Embeddings
sentence-transformers>=2.2.2

# Vector database
chromadb>=0.4.22

# # Local LLM (Ollama integration)
# ollama>=0.1.6

# Utilities
tqdm>=4.65.0


streamlit>=1.30.0


newsapi-python